aksd: Extract shared components from CreateAKSProject
