+++ plugins/aks-desktop/src/components/CreateAKSProject/types.ts
@@ -138,37 +138,12 @@ export interface BreadcrumbProps {
   onStepClick: (step: number) => void;
 }
 
-export interface FormFieldProps {
-  label: string;
-  value: string | number;
-  onChange: (value: string | number) => void;
-  type?: 'text' | 'email' | 'number' | 'textarea';
-  multiline?: boolean;
-  rows?: number;
-  placeholder?: string;
-  error?: boolean;
-  helperText?: string;
-  disabled?: boolean;
-  required?: boolean;
-  startAdornment?: React.ReactNode;
-  endAdornment?: React.ReactNode;
-  /** Ref for the input elemenet of this field */
-  inputRef?: React.Ref<HTMLInputElement>;
-}
-
 export interface ValidationAlertProps {
   type: 'error' | 'warning' | 'success' | 'info';
   message: string | React.ReactNode;
   onClose?: () => void;
   action?: React.ReactNode;
   show?: boolean;
-}
-
-export interface ResourceCardProps {
-  title: string;
-  icon: string;
-  iconColor: string;
-  children: React.ReactNode;
 }
 
 // Validation result types
@@ -233,16 +208,3 @@ export function mapUIRoleToAzureRole(uiRole: string): string {
 
   return roleMap[uiRole] || uiRole; // Fallback to original if not found
 }
-
-// Networking policy options
-export const INGRESS_OPTIONS = [
-  { value: 'AllowSameNamespace', label: 'Allow traffic within same namespace' },
-  { value: 'AllowAll', label: 'Allow all traffic' },
-  { value: 'DenyAll', label: 'Deny all traffic' },
-] as const;
-
-export const EGRESS_OPTIONS = [
-  { value: 'AllowAll', label: 'Allow all traffic' },
-  { value: 'AllowSameNamespace', label: 'Allow traffic within same namespace' },
-  { value: 'DenyAll', label: 'Deny all traffic' },
-] as const;
